src/main/java/badas/dev/gerenciamento/de/estoque/service/ProdutoService.java
package badas.dev.gerenciamento.de.estoque.service;

import badas.dev.gerenciamento.de.estoque.model.Produto;
import badas.dev.gerenciamento.de.estoque.repository.ProdutoRepositorio;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class ProdutoService {

    @Autowired
    private ProdutoRepositorio produtoRepositorio;

    public Produto addProduto(Produto produto) {
        return produtoRepositorio.save(produto);
    }
}
